File storage before pause
