+++ traits/Settings.php
@@ -3,19 +3,17 @@
  * Modularised settings across the plugin, in the WordPress way
  */
 
-namespace MOJElasticSearch;
-
-use function ElasticPress\Utils\is_indexing;
+namespace MOJElasticSearch\Traits;
 
 trait Settings
 {
     public function createSections($group)
     {
-        if (empty($group) || !isset(Admin::$sections[$group]) || !is_array(Admin::$sections[$group])) {
+        if (empty($group) || !isset(Page::$sections[$group]) || !is_array(Page::$sections[$group])) {
             return;
         }
 
-        foreach (Admin::$sections[$group] as $section) {
+        foreach (Page::$sections[$group] as $section) {
             add_settings_section(
                 $this->prefix . '_' . $section['id'],
                 __($section['title'], $this->text_domain),
